move loading lottie compose
